+Enabled bonded elements by name, not by molecule as it was handled in agglutination. CellBase is modified in order to acept a name reference (inclution in simulation not mandatory)
Note: which it lasts is check whether the pathogen reference is type of
Phage. This could be done in the loading functions (so that it was
possible to throw and exception before starting the simulation. Dont
stop the simulation if that occurs, just give a message and if there are
not any bold well declared, infection module must be desactivated)

Extra: could be defined multiple host whithin a single bond tag????

This closes #8
This closes #9
This closes #10